[FLINK-39499][filesystems/tests] Switch from minio to seaweedFS in flink-filesystems
Following the recent deprecation of minIO
