+++ flink-filesystems/flink-s3-fs-hadoop/src/test/java/org/apache/flink/fs/s3hadoop/HAApplicationRunOnHadoopS3FileSystemITCase.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.fs.s3hadoop;
 
-import org.apache.flink.fs.s3.common.HAApplicationRunOnMinioS3StoreITCase;
+import org.apache.flink.fs.s3.common.HAApplicationRunOnSeaweedFsS3StoreITCase;
 
-/** Runs the {@link HAApplicationRunOnMinioS3StoreITCase} on the Hadoop S3 file system. */
-class HAApplicationRunOnHadoopS3FileSystemITCase extends HAApplicationRunOnMinioS3StoreITCase {}
+/** Runs the {@link HAApplicationRunOnSeaweedFsS3StoreITCase} on the Hadoop S3 file system. */
+class HAApplicationRunOnHadoopS3FileSystemITCase extends HAApplicationRunOnSeaweedFsS3StoreITCase {}
